main,refactor: make sort.h main part private
