+++ win32/ctags_vs2013.vcxproj
@@ -291,7 +291,7 @@
     <ClInclude Include="..\main\routines.h" />
     <ClInclude Include="..\main\routines_p.h" />
     <ClInclude Include="..\main\selectors.h" />
-    <ClInclude Include="..\main\sort.h" />
+    <ClInclude Include="..\main\sort_p.h" />
     <ClInclude Include="..\main\strlist.h" />
     <ClInclude Include="..\main\subparser.h" />
     <ClInclude Include="..\main\tokeninfo.h" />
