+++ win32/ctags_vs2013.vcxproj.filters
@@ -626,7 +626,7 @@
     <ClInclude Include="..\main\selectors.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="..\main\sort.h">
+    <ClInclude Include="..\main\sort_p.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="..\main\strlist.h">
